requirements.txt
click==8.1.3
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
imutils==0.5.4
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.1
opencv-contrib-python
opencv-contrib-python-headless
pyserial
pyttsx3
cryptography
pandastable